Tests: Modifications to main test

+++ arc/mainTest.py
@@ -41,6 +41,12 @@ class TestARC(unittest.TestCase):
                           'lennard_jones': False,
                           'bde': True,
                           }
+        projects = ['arc_project_for_testing_delete_after_usage_test_from_dict',
+                    'arc_model_chemistry_test', 'arc_test', 'test', 'unit_test_specific_job', 'wrong']
+        for project in projects:
+            project_directory = os.path.join(ARC_PATH, 'Projects', project)
+            if os.path.isdir(project_directory):
+                shutil.rmtree(project_directory, ignore_errors=True)
 
     def test_as_dict(self):
         """Test the as_dict() method of ARC"""
@@ -424,7 +430,8 @@ class TestARC(unittest.TestCase):
                     'arc_model_chemistry_test', 'arc_test', 'test', 'unit_test_specific_job', 'wrong']
         for project in projects:
             project_directory = os.path.join(ARC_PATH, 'Projects', project)
-            shutil.rmtree(project_directory, ignore_errors=True)
+            if os.path.isdir(project_directory):
+                shutil.rmtree(project_directory, ignore_errors=True)
 
 
 if __name__ == '__main__':
